Fixing 2018 omission
The code I submitted on the due date omitted 2018 from the final dataframe because I messed up the range in the final merging function. I need to update the repository because I am sharing the project with some people for advocacy purposes, but I understand that my project grade is based on what I submitted earlier.